+++ MelonStartScreen/UI/UIStyleValues.cs
@@ -20,8 +20,7 @@ namespace MelonLoader.MelonStartScreen.UI
         internal static void Init()
         {
             TextFont = Resources.GetBuiltinResource<Font>("Arial.ttf");
-            BackgroundTexture = UIUtils.CreateColorTexture(UICustomization.General.BackgroundColor);
-            BackgroundImage = LoadImage("Background");
+            BackgroundTexture = UIUtils.CreateColorTexture(UICustomization.Background.SolidColor);
 
             if (UICustomization.ProgressBar.Enabled)
             {
@@ -29,39 +28,44 @@ namespace MelonLoader.MelonStartScreen.UI
                 ProgressBarOuterTexture = UIUtils.CreateColorTexture(UICustomization.ProgressBar.OuterColor);
             }
 
-            if (UICustomization.Logo.Enabled)
+            if (UICustomization.Background.ScanForCustomImage)
+                BackgroundImage = LoadImage("Background", UICustomization.Background.Filter);
+
+            if (UICustomization.LogoImage.Enabled)
             {
-                LogoImage = LoadImage("Logo");
+                if (UICustomization.LogoImage.ScanForCustomImage)
+                    LogoImage = LoadImage("Logo", UICustomization.LogoImage.Filter);
                 if (LogoImage == null)
                     LogoImage = new Image(
                     (MelonLaunchOptions.Console.Mode == MelonLaunchOptions.Console.DisplayMode.LEMON)
                         ? Properties.Resources.Logo_Lemon
-                        : Properties.Resources.Logo_Melon);
+                        : Properties.Resources.Logo_Melon, UICustomization.LogoImage.Filter);
             }
 
-            if (UICustomization.Animation.Enabled)
+            if (UICustomization.LoadingImage.Enabled)
             {
-                LoadingImage = LoadImage("Loading");
+                if (UICustomization.LoadingImage.ScanForCustomImage)
+                    LoadingImage = LoadImage("Loading", UICustomization.LoadingImage.Filter);
                 if (LoadingImage == null)
                     LoadingImage = new AnimatedImage(
                     (MelonLaunchOptions.Console.Mode == MelonLaunchOptions.Console.DisplayMode.LEMON)
                         ? Properties.Resources.Loading_Lemon
-                        : Properties.Resources.Loading_Melon);
+                        : Properties.Resources.Loading_Melon, UICustomization.LoadingImage.Filter);
             }
         }
 
-        private static Image LoadImage(string filename)
+        private static Image LoadImage(string filename, FilterMode filterMode = FilterMode.Bilinear)
         {
             string filepath = ScanForFile(filename);
             if (string.IsNullOrEmpty(filepath))
                 return null;
             string fileext = Path.GetExtension(filepath).ToLowerInvariant();
             if (fileext.Equals(".gif"))
-                return new AnimatedImage(filepath);
+                return new AnimatedImage(filepath, filterMode);
             if (fileext.Equals(".png")
                 || fileext.Equals(".jpg")
                 || fileext.Equals(".jpeg"))
-                return new Image(filepath);
+                return new Image(filepath, filterMode);
             return null;
         }